# backend/src/core/auth/dto/reset-password-request.dto.ts
import { AuthAdmin } from '@generated/prisma-client';
import { IsNotEmpty, IsString } from 'class-validator';
import { IsOneTimeToken } from '../decorators';
import { ApiProperty } from '@nestjs/swagger';

export class ResetPasswordRequestDto implements Pick<AuthAdmin, 'password'> {
  @ApiProperty({ description: 'One time token' })
  @IsString()
  @IsNotEmpty()
  @IsOneTimeToken()
  readonly token!: Base64URLString;

  // password strength is checked in the service
  @ApiProperty({ format: 'password' })
  @IsString()
  @IsNotEmpty()
  readonly password!: string;
}